Reorganize file and fmt
