+++ src/typemap/src/lib.rs
@@ -1,7 +1,7 @@
-pub mod fs_conv;
-pub mod syscall_conv;
-pub mod type_conv;
+pub mod cage_helpers;
+pub mod datatype_conversion;
+pub mod path_conversion;
 
-pub use fs_conv::*;
-pub use syscall_conv::*;
-pub use type_conv::*;
+pub use cage_helpers::*;
+pub use datatype_conversion::*;
+pub use path_conversion::*;
